refactor: movendo popup de deletar para molecules
