+++ src/components/molecules/DeletePostPopUp/styles.ts
